pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pitch-deck-analyzer"
version = "0.1.0"
description = "AI-Driven Pitch Deck Analysis Platform using CrewAI"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "crewai>=0.11.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.0",
    "groq>=0.4.0",
    "langchain-groq>=0.1.0",
    "python-pptx>=0.6.23",
    "PyPDF2>=3.0.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "python-docx>=1.1.0",
    "langchain>=0.1.0",
    "pyyaml>=6.0.1"
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0"
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 